std.os.linux: make dev_t u32
glibc and musl define dev_t as u64, but in the kernel it is u32.

see https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/linux/types.h?id=v7.0#n17

+++ lib/std/os/linux.zig
@@ -105,7 +105,7 @@ pub const user_desc = arch_bits.user_desc;
 
 pub const blkcnt_t = u64;
 pub const blksize_t = u32;
-pub const dev_t = u64;
+pub const dev_t = u32;
 pub const ino_t = u64;
 pub const mode_t = u32;
 pub const nlink_t = u32;
